fix test data

+++ tests/test_any.nim
@@ -69,7 +69,7 @@ test "negative int":
   check cmp(anyInt(-128), "d080")
   check cmp(anyInt(-32768), "d18000")
   check cmp(anyInt(-65536), "d2FFFF0000")
-  check cmp(anyInt(int64(low(int32)-1)), "D3FFFFFFFF7FFFFFFF")
+  check cmp(anyInt(int64(low(int32))-1), "D3FFFFFFFF7FFFFFFF")
 
 test "float":
   check cmp(anyFloat(0.0'f64), "cb0000000000000000")
@@ -139,11 +139,11 @@ test "ordinal 16 bit":
 
 test "ordinal 32 bit":
   let uu = [low(int32), low(int32)+1, low(int32)+2, high(int32)-1,
-    high(int32)-2, low(int8)-2, low(int8)-1, low(int8), low(int8)+1,
-    low(int8)+2, low(int16)-2, low(int16)-1, low(int16), low(int16)+1,
-    low(int16)+2, high(int8)-2, high(int8)-1, high(int8), high(int8)+1,
-    high(int8)+2, high(int16)-2, high(int16)-1, high(int16), high(int16)+1,
-    high(int16)+2,high(int32)]
+    high(int32)-2, int32(low(int8))-2, int32(low(int8))-1, low(int8), low(int8)+1,
+    low(int8)+2, int32(low(int16))-2, int32(low(int16))-1, low(int16), low(int16)+1,
+    low(int16)+2, high(int8)-2, high(int8)-1, high(int8), int32(high(int8))+1,
+    int32(high(int8))+2, high(int16)-2, high(int16)-1, high(int16), int32(high(int16))+1,
+    int32(high(int16))+2,high(int32)]
 
   let vv = [low(uint32), low(uint32)+1, low(uint32)+2, high(uint32), high(uint32)-1,
     high(uint32)-2, low(uint8), low(uint8)+1, low(uint8)+2, low(uint16)+1,
@@ -174,13 +174,13 @@ test "ordinal 32 bit":
 
 test "ordinal 64 bit":
   let uu = [high(int64), low(int32), low(int32)+1, low(int32)+2, high(int32)-1,
-    high(int32)-2, low(int8)-2, low(int8)-1, low(int8), low(int8)+1,
-    low(int8)+2, low(int16)-2, low(int16)-1, low(int16), low(int16)+1,
-    low(int16)+2, high(int8)-2, high(int8)-1, high(int8), high(int8)+1,
-    high(int8)+2, high(int16)-2, high(int16)-1, high(int16), high(int16)+1,
-    high(int16)+2,high(int32), low(int64)+1, low(int64)+2, low(int64),
+    high(int32)-2, int64(low(int8))-2, int64(low(int8))-1, low(int8), low(int8)+1,
+    low(int8)+2, int64(low(int16))-2, int64(low(int16))-1, low(int16), low(int16)+1,
+    low(int16)+2, high(int8)-2, high(int8)-1, high(int8), int64(high(int8))+1,
+    int64(high(int8))+2, high(int16)-2, high(int16)-1, high(int16), int64(high(int16))+1,
+    int64(high(int16))+2,high(int32), low(int64)+1, low(int64)+2, low(int64),
     high(int64)-1, high(int64)-2, low(int64),low(int64)+1,low(int64)+2,
-    low(int32)-1,low(int32)-2]
+    int64(low(int32))-1,int64(low(int32))-2]
 
   var s = initMsgStream()
   for i in uu: s.pack(i)
@@ -238,14 +238,14 @@ test "map copy and `in` operator":
   check anyString("abc") in b
   var c = anyArray(anyString("apple"))
   check anyString("apple") in c
-  
+
 test "bin and ext":
   const extType = 0xCE'i8
   var bin = anyBin("binary data")
   var ext = anyExt("ext oi...", extType)
   var arr = anyArray(bin, ext)
   var msg = arr.fromAny()
-  
+
   var a_arr = toAny(msg)
   check a_arr[0].kind == msgBin
   check a_arr[1].kind == msgExt
@@ -253,4 +253,3 @@ test "bin and ext":
   check a_arr[1].extType == extType
   check a_arr[1].extData == "ext oi..."
 
-
